move 'brainobjects', 'models', 'inputs' into brainpy/building
