+++ docs/apis/building.rst
@@ -1,0 +1,15 @@
+``brainpy.building`` module
+=============================
+
+.. currentmodule:: brainpy.building
+.. automodule:: brainpy.building
+
+
+.. toctree::
+    :maxdepth: 1
+
+    auto/building/brainobjects
+    auto/building/models
+    auto/building/connect
+    auto/building/inputs
+
