+++ docs/apis/simulation.rst
@@ -8,10 +8,6 @@
 .. toctree::
     :maxdepth: 1
 
-    auto/simulation/brainobjects
-    auto/simulation/models
-    auto/simulation/connect
-    auto/simulation/inputs
     auto/simulation/measure
     auto/simulation/monitor
     auto/simulation/runner
